Extract JsonObjectParser to top-level
