+++ parsing-utilities/src/main/java/gov/nasa/jpl/aerie/json/ProductParsers.java
@@ -18,33 +18,6 @@ public abstract class ProductParsers {
 
   public static final EmptyProductParser productP = new EmptyProductParser();
 
-  public interface JsonObjectParser<T> extends JsonParser<T> {
-    @Override JsonObject unparse(final T value);
-
-    @Override default <S> JsonObjectParser<S> map(final Iso<T, S> transform) {
-      Objects.requireNonNull(transform);
-
-      final var self = this;
-
-      return new JsonObjectParser<>() {
-        @Override
-        public JsonObject getSchema(final SchemaCache anchors) {
-          return self.getSchema(anchors);
-        }
-
-        @Override
-        public JsonParseResult<S> parse(final JsonValue json) {
-          return self.parse(json).mapSuccess(transform::from);
-        }
-
-        @Override
-        public JsonObject unparse(final S value) {
-          return self.unparse(transform.to(value));
-        }
-      };
-    }
-  }
-
 
   public static final class EmptyProductParser implements JsonObjectParser<Unit> {
     private EmptyProductParser() {}
